chat-bot/src/requirements.txt
pymupdf
tqdm
faiss-cpu
numpy
langchain-experimental
langchain-huggingface
sentence-transformers
colorlog
pydantic-settings
fastapi
uvicorn
langdetect
cohere